Rewrote Tomi's fix as the equivalent:
use open ':locale'

which is at the same time locale-sensitive.

NB! This means that Perl will only treat std IO as UTF-8 iff the locale is a UTF-8 locale. Since that is our standard encoding everywhere, it shouldn't be a problem.

Also removed the whole commented line that was the real cause of bug #211,