Added a simple perl script that uses Text::Brew to calculate the editing distance between a pair of words.
INPUT:
(a file with) lines as follows:

string_with_spelling_error<TAB>correct_string

OUTPUT:
lines of the following format:

edit_distance<TAB>error<TAB>correct<TAB>(sequence of edit operations to go from error to correct)

The script requires that Text::Brew is available to perl (e.g. in one of the PERLLIB dirs). Text::Brew can be downloaded and installed from CPAN (try 'cpan' on the command line).

Known bugs:

Text::Brew is NOT Unicode aware, and gives wrong results for non-ASCII chars, counting the number of 8-bit bytes to change when calculating the editing distance (that is, each SaÌmi char will give a distance of 2, because they occupy 2 bytes in the UTF-8 sequence). This gives a *larger* Brew/Levenshtein distance than expected, but the result is still quite useful. Feel free to try to correct it - the whole Text::Brew module is pure perl;-)