Changed the script to output proper XInclude's instead of just ID strings in the ref to lang entries in the termcenter file.
This change increases the size of the file quite a lot, but it simplifies processing in the db *tremendeously*. Another drawback of this change is that it is harder to edit the raw XML manually - perhaps we should have some sort of post-processor/pre-commit checks and clean-ups for manual editing? When editing the files in eXist/risten.no, there's no difference, the references are handled automatically.