Made the script work, provided the user adds one of these lines to his/her .bashrc: export GT_HOME=/Users/<username>/ export GT_HOME=/Users/<username>/Documents/ (according to whether the gt/ catalogue is in the home dir or in Documents).
The script now works with plain text input files, it can be tweaked in different ways.

I also added a plan for replacing this script, and the family of cealkka, sme-dis.sh, smj-multi.sh etc. with one script with several options. The options, their default values, etc. were sketched.

With this sketch the script leaves the realm of my scripting competence, and I hope for input from computer-literates.