Simple perl script to pretty-print xml using XML::Twig. Input can be a file or STDIN.
NB! The whole file is read into memory before being printed out, so be careful with big files! It has been tested on a 5 Mb file without problems, and given that all of us have at least 1Gb of RAM, it shouldn't be a big concern. But you have been warned:-)