Script to run a text through MS Word's spell checker without user interaction, and gathering the result in a separate file.
From the AppleScript file:

The input file is expected to have one word on each line. It is possible that the script would work without this limitation, but that is not tested.

OUTPUT:

A tab-separated file of the following scheme:
Orig	SpellerCategory	Suggestions

where:

Orig = the original string in the input file (ie the word to be spell-checked)
SpellerCategory = either one of:
	- SplCor (correct spelling according to the speller)
	- SplErr (spelling error according to the speller
	- CapErr (capitalization error, don't know exactly what this means)
Suggestions = list of suggestions given by the speller, potentially empty