The previous solution of storing the output from the Word speller in one, large variable, and then sending it all to STOUT when finished, didn't work - we very soon reached the upper limit for variables in AppleScript.
The script and the Makefile is now changed to print out the result of each word being checked as we move along.

Please note that the content is written to a temp. file in UTF-16 (that is what AppleScript (or MS Word?) gives), which is in the end converted back to UTF-8 in the final result file. This said just in case somebody wants to peek into the generated file before the spell checking is ready.