Added a set of other, potentially interesting languages to spell check.
Since this script isn't really dependent on any binary files beyond what is delivered with MS Word, we can use this AppleScript to test any of the MS spellers. This can be usefull for a number of different purposes.

The list of languages can be extended by just providing a mapping from language codes to MS language identifier strings. A link to documentation for these strings will be provided in the upcoming documentation for the AppleScript/Word test bench.