â¢ made the script aware of different versions of Word (2004 and 2008) - they can coexist on the same computer, and the specified one will be called â¢ moved the actual spelling checking and related processing to a separate subroutine to be able to handle the different Word versions
I had to make two variants of the same subroutine, with the only real difference being which application i "tell". But since there appears to be other differences in the AS implementations as well, it is probably ok.

Word 2008 is not used unless explicitly asked for - the default is Word 2004. This should remain so until we get proofing tools for Office 2008.