First version of a setup script to be run directly after the initial checkout of gtsvn.
Presently it only defines an env. variable

$GTHOME

and adds the path to the gt/script/ directory to the $PATH env. variable.

Future versions will add checks for software, setup the lookup shortcuts, etc.

The setup script should work in bash and a couple of C shells, and should work on most Unix-like systems (but not on Windows). C shell functionality is not yet tested.

Paths containing spaces are ok.

This version is NOT yet ready for general consumption, but interested earlybirds are welcome to enhance, test and give feedback.