Several changes to make the script more robust: - first we make a backup of the original file - then we make a copy of the original file, filtering out the init.sh line if found - all changes are made on the copy - all changes are displayed to the user before a final confirmation - only if the user answers YES in this last case the changes are copied back to the original file
This should ensure that nothing is broken in case of a power out situation - in the worst case the system goes down in the middle of the copy-back process, in case you still have the backup to fall back to.

As a consequence of these changes, some of the dialog texts are not completely meaningful anymore, and the whole dialog structure needs a makeover. But the scripts are functional.