Fixed a bug where the common.xsl file would only be imported from the location:
/usr/local/share/corp/bin/common.xsl

Now, with the help of the helper xsl stylesheet preprocxsl.xsl, the file-specific stylesheet is preprocessed to import the $GTHOME/gt/script/common.xsl stylesheet instead. This is done by creating a temporary xsl file where the import statement is replaced with a working one, which is then used in place of the source xsl file.

Within XSLT 1.0 this is the simplest solution, since there is no means for indirection of the import href, and relative refs would not work.

The bug caused local corpus conversion to fail, ie it would only work on victorio. Now it works everywhere.