Removed the alias sortcyr, it did not work (it sorted strings of variable length separately).
Added the alias
alias sortr='LC_ALL="ru" sort'
(thanks to Cip), and called it sortr, since what
we use is the Russian locale. Whether this is the best
solution for, say Kildin, remains to be seen. In any case
it works for the А-Я span. Now we thus can compare Kildin lists.